fix: add workspace file in spa_tmp forder to satisfy pnpm 11 requrements, so packages from the custom folders can  execute scripts
AdminForth/1821/agent-plugins-doesn't-work-out

+++ dev-demo/resources/cars_resources/carsResourseTemplate.ts
@@ -134,7 +134,7 @@ export default function carsResourseTemplate(resourceId: string, dataSource: Car
       {
         name: 'mileage',
         allowMinMaxQuery: true,
-        type: AdminForthDataTypes.FLOAT,
+        type: AdminForthDataTypes.DECIMAL,
       },
       {
         name: 'color',
